+++ lib/dealog_backoffice/release.ex
@@ -1,6 +1,13 @@
 defmodule DealogBackoffice.Release do
+  @moduledoc """
+  Functions for running one-off tasks on a released version of the backoffice.
+  """
+
   @app :dealog_backoffice
 
+  @doc """
+  Ensure event store is setup and initialized.
+  """
   def init_event_store do
     load_app()
 
@@ -10,6 +17,9 @@ defmodule DealogBackoffice.Release do
     :ok = EventStore.Tasks.Init.exec(DealogBackoffice.EventStore, config, [])
   end
 
+  @doc """
+  Apply database migrations.
+  """
   def migrate do
     load_app()
 
@@ -18,9 +28,24 @@ defmodule DealogBackoffice.Release do
     end
   end
 
+  @doc """
+  Rollback database migrations.
+  """
   def rollback(repo, version) do
     load_app()
     {:ok, _, _} = Ecto.Migrator.with_repo(repo, &Ecto.Migrator.run(&1, :down, to: version))
+  end
+
+  @doc """
+  Import the list of administrative areas.
+  """
+  def import_administrative_areas do
+    load_app()
+    {:ok, _} = Application.ensure_all_started(:dealog_backoffice)
+
+    :dealog_backoffice
+    |> Application.app_dir(["priv", "repo", "administrative_areas"])
+    |> DealogBackoffice.Importer.AdministrativeAreas.import_all_from()
   end
 
   defp repos do