+++ priv/repo/import_administrative_areas.exs
@@ -1,52 +1,9 @@
-defmodule DealogBackoffice.Importer.AdministrativeAreas do
-  @moduledoc """
-  Importer for CSV based administrative areas.
-  """
-
-  alias DealogBackoffice.Repo
-
-  def import_all_from(path) do
-    collect_import_files(path)
-    |> Enum.reduce(0, fn file, total ->
-      {amount, _} = import_file(path, file)
-      total + amount
-    end)
-  end
-
-  defp collect_import_files(path) do
-    File.ls!(path)
-    |> Enum.filter(&(Path.extname(&1) == ".csv"))
-    |> Enum.sort()
-  end
-
-  defp import_file(path, file_name) do
-    full_path = Path.join(path, file_name)
-    ags = Path.rootname(file_name)
-    IO.puts("\tImporting ags #{ags}")
-
-    entries =
-      File.stream!(full_path)
-      |> CSV.decode!(strip_fields: true, headers: true)
-      |> Enum.map(
-        &%{
-          ags: &1["ags"],
-          type_label: &1["bez"],
-          type: &1["type"],
-          name: &1["gen"],
-          parent_ags: &1["parent_ags"],
-          inserted_at: NaiveDateTime.utc_now(),
-          updated_at: NaiveDateTime.utc_now()
-        }
-      )
-
-    Repo.insert_all("administrative_areas", entries, on_conflict: :nothing)
-  end
-end
+alias DealogBackoffice.Importer.AdministrativeAreas
 
 path = Path.expand(__DIR__ <> "/administrative_areas")
 
 IO.puts("Importing administrative areas from #{path}")
 
-amount = DealogBackoffice.Importer.AdministrativeAreas.import_all_from(path)
+amount = AdministrativeAreas.import_all_from(path)
 
 IO.puts("Imported #{amount} new administrative areas")