+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- `CNG` Import of administrative areas is running on container start (until there is support for one-off tasks)
 - `ADD` Administrative areas can be imported
 - `FIX` The Kafka host can be configured via an env var
 - `ADD` Published message details can be viewed